refactor: Rename zerodw package
